Add GlobalState component to App.js and fetch weather data in GlobalState component
