+++ WebApp/src/App.tsx
@@ -1,9 +1,12 @@
 import React from 'react';
 import Header from './components/header/Header';
+import GlobalState from './context/GlobalState';
 
 function App() {
   return (
-    <div className='h-screen'>{Header()}</div>
+    <GlobalState>
+      <div className='h-screen'>{Header()}</div>
+    </GlobalState>
   );
 }
